+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 1.1.0
+
++ [#16](https://github.com/luyadev/luya-generic/issues/16) Add list type and CSS classes to `FileListBlock`
++ Updated Composer and `luyadev/luya-testsuite`
+
 ## 1.0.6 (11. February 2020)
 
 + Small french translation updates.